Show queried labels in frontend
