+++ frontend/src/app/annotate/annotation-input/problem-details/problem-labels/problem-labels.component.ts
@@ -1,10 +1,9 @@
-import { Component, computed, input, OnInit } from '@angular/core';
-import { ProblemLabel } from '../problem-details.component';
+import { Component, computed, input } from '@angular/core';
 import { faSearch } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { NgbTooltipModule, NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { FormControl } from '@angular/forms';
 import { ManageLabelsModalComponent } from './manage-labels-modal/manage-labels-modal.component';
+import { ProblemLabel } from '@/types';
 
 @Component({
     selector: 'la-problem-labels',
